+++ internal/app/api/infrastructure/database/agent.go
@@ -1,4 +1,4 @@
-package infrastructure
+package database
 
 import (
 	"context"
