+++ internal/app/api/infrastructure/database/policy.go
@@ -1,4 +1,4 @@
-package infrastructure
+package database
 
 import (
 	"context"
