+++ internal/app/api/infrastructure/database/user_token.go
@@ -1,4 +1,4 @@
-package infrastructure
+package database
 
 import (
 	"context"
